requirements.txt
alabaster==0.7.10
amqp==2.2.2
appnope==0.1.0
arrow==0.10.0
asn1crypto==0.22.0
aspy.yaml==1.0.0
astroid==1.4.9
Babel==2.4.0
bandit==1.4.0
billiard==3.5.0.3
binaryornot==0.4.4
bleach==1.5.0
boltons==16.1.1
brotlipy==0.7.0
cached-property==1.3.1
celery==4.1.0
certifi==2017.11.5
cffi==1.10.0
chardet==3.0.4
click==6.7
conda==4.3.22
cookiecutter==1.5.1
coverage==4.4.2
cryptography==1.8.1
decorator==4.0.11
dj-database-url==0.4.2
Django==1.11.7
django-js-reverse==0.7.3
django-log-request-id==1.3.2
django-model-utils==3.0.0
django-naomi==0.8
django-webpack-loader==0.5.0
docutils==0.13.1
dodgy==0.1.9
dparse==0.2.1
entrypoints==0.2.3
fixmydjango==0.3.1
future==0.16.0
gitdb2==2.0.3
GitPython==2.1.7
gunicorn==19.7.1
html5lib==0.999
identify==1.0.7
idna==2.5
imagesize==0.7.1
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
isort==4.2.15
jedi==0.10.2
Jinja2==2.9.6
jinja2-time==0.2.0
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
kombu==4.1.0
lazy-object-proxy==1.3.1
MarkupSafe==0.23
mccabe==0.6.1
mistune==0.7.4
model-mommy==1.4.0
nbconvert==5.2.1
nbformat==4.3.0
nltk==3.2.4
nodeenv==1.2.0
numpy==1.13.1
numpydoc==0.6.0
olefile==0.44
opbeat==3.5.3
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.1
pbr==3.1.1
pep8==1.7.0
pep8-naming==0.4.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
poyo==0.4.1
pre-commit==1.4.1
prompt-toolkit==1.0.14
prospector==0.12.7
psutil==5.2.2
psycopg2==2.7.3.2
ptyprocess==0.5.1
pycodestyle==2.0.0
pycosat==0.6.2
pycparser==2.17
pydocstyle==2.1.1
pyflakes==1.5.0
pygame==1.9.3
Pygments==2.2.0
pylint==1.6.4
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.7.2
pylint-flask==0.5
pylint-plugin-utils==0.2.6
pyOpenSSL==17.0.0
pyparsing==2.1.4
python-dateutil==2.6.0
python-decouple==3.1
pytz==2017.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.0
QtPy==1.2.1
redis==2.10.6
requests==2.14.2
requirements==0.1
requirements-detector==0.5.2
rope-py3k==0.9.4.post1
safety==1.6.1
scikit-learn==0.18.2
setoptconf==0.2.0
simplegeneric==0.8.1
six==1.10.0
smmap2==2.0.3
snowballstemmer==1.2.1
Sphinx==1.6.2
sphinxcontrib-websupport==1.0.1
spyder==3.1.4
stevedore==1.27.1
termcolor==1.1.0
testpath==0.3
tornado==4.5.1
traitlets==4.3.2
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
vulture==0.26
wcwidth==0.1.7
webcolors==1.7
whichcraft==0.4.1
whitenoise==3.3.1
wrapt==1.10.10
xlrd==1.1.0
XlsxWriter==1.0.2
